Add victim spawn points to map prefabs
Added FloodVictim and EarthquakeVictim GameObjects with multiple spawn point transforms to map1.prefab and map2.prefab. This enables dynamic spawning of victims for flood and earthquake scenarios in the maps using the updated VictimSpawner.cs. Also updated parent-child relationships for fence objects to improve scene organization.
Deleted the FloodMapSpawner.cs as it is  not used.